backend/requirements.txt
absl-py==2.2.2
aioice==0.10.1
aiortc==1.11.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
asyncio==3.4.3
attrs==25.3.0
av==14.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
jax==0.6.0
jaxlib==0.6.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.9.2
kiwisolver==1.4.8
langchain-core==0.3.56
langgraph==0.3.34
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.37
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
namex==0.0.9
ndg-httpsclient==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==4.25.7
pusher==3.3.3
pusherclient==0.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.2
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
websockets==15.0.1